CIRC-4909 Fix formatting on reconstitute page
The main issue was the indented code blocks, which were done a
different way in the pre-Hugo regime.

Remove the need to get the node ID, because we can just use the normal
start script with some extra args. The start script already knows how
to get the node ID.